modificacion2 del archivo bd_ramos

+++ bd_ramos.sql
@@ -1,8 +1,8 @@
 create table Curso (
     curso_id serial not null,
     curso_nombre VARCHAR(100) not null,
-    curso_periodos INTEGER(2) not null,
-    fase_modulo INTEGER(2) not null,
+    curso_periodos INTEGER not null,
+    fase_modulo INTEGER not null,
     Instructor_nombre VARCHAR(100) not null,
     creditos_curso DECIMAL(3,2),
     primary key (curso_id)
